requirements.txt
Click==7.0
Flask==1.0.2
ImageHash==4.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.2
Pillow==5.3.0
PyWavelets==1.0.1
scipy==1.1.0
six==1.11.0
tinydb==3.11.1
Werkzeug==0.14.1
